chore(settings): adds zbus proxies from common zbus client

+++ shell/status-bar/src/modules/battery/service.rs
@@ -33,6 +33,6 @@ impl BatteryService {
             )),
         };
 
-        Ok((capacity, status))
+        Ok((capacity as u8, status))
     }
 }
